requirements.txt
absl-py==1.4.0
argcomplete==2.0.0
asgiref==3.6.0
astunparse==1.6.3
Beaker==1.12.1
beautifulsoup4==4.12.2
blivet==3.7.1
blivet-gui==2.4.1
Brlapi==0.8.4
cachetools==5.3.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==37.0.2
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.8.0
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
dnspython==2.3.0
fedora-third-party==0.10
file-magic==0.4.0
flatbuffers==23.3.3
fros==1.1
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.17.1
grpcio==1.54.0
h5py==3.8.0
humanize==3.13.1
idna==3.4
imutils==0.5.4
jax==0.4.8
keras==2.12.0
langtable==0.0.62
libclang==16.0.0
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
nftables==0.1
numpy==1.23.5
olefile==0.46
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==9.4.0
ply==3.11
productmd==1.35
protobuf==4.22.3
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1-modules==0.3.0
pycairo==1.23.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGObject==3.44.1
pykickstart==3.47
pyOpenSSL==21.0.0
pyparted==3.12.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
pytz==2023.3
pyudev==0.24.0
pyxdg==0.27
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.18.1
rsa==4.9
scipy==1.10.1
selinux==3.5
sepolicy==3.5
setools==4.4.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.5.1
soupsieve==2.4.1
sqlparse==0.4.3
systemd-python==235
Tempita==0.5.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
urllib3==1.26.15
wrapt==1.14.1
